wip! authz

+++ src/endpoints/sources.js
@@ -11,7 +11,7 @@ const stream = require("stream");
 const {promisify} = require("util");
 const zlib = require("zlib");
 
-const {AuthzDenied} = require("../exceptions");
+const authz = require("../authz");
 const {contentTypesProvided, contentTypesConsumed} = require("../negotiate");
 const {fetch, Request} = require("../fetch");
 const sources = require("../sources");
@@ -44,9 +44,7 @@ const setSource = (sourceName, argsExtractor = (req) => []) => (req, res, next) 
 
   res.vary("Accept");
 
-  if (!source.visibleToUser(req.user)) {
-    throw new AuthzDenied();
-  }
+  authz.assertAuthorized(req.user, authz.actions.Read, source);
 
   req.context.source = source;
   return next();
@@ -122,6 +120,8 @@ const canonicalizeDataset = (pathBuilder) => (req, res, next) => {
  * @type {expressMiddlewareAsync}
  */
 const ifDatasetExists = async (req, res, next) => {
+  authz.assertAuthorized(req.user, authz.actions.Read, req.context.dataset);
+
   if (!(await req.context.dataset.exists())) throw new NotFound();
   return next();
 };
@@ -208,6 +208,8 @@ const putDataset = contentTypesConsumed([
 const deleteResource = resourceExtractor => async (req, res) => {
   const resource = resourceExtractor(req);
 
+  authz.assertAuthorized(req.user, authz.actions.Write, resource);
+
   const method = "DELETE";
   const upstreamUrls = await Promise.all(resource.subresources().map(s => s.url(method)));
   const upstreamResponses = await Promise.all(upstreamUrls.map(url => fetch(url, {method})));
@@ -249,6 +251,8 @@ const setNarrative = (pathExtractor) => (req, res, next) => {
  * @type {expressMiddlewareAsync}
  */
 const ifNarrativeExists = async (req, res, next) => {
+  authz.assertAuthorized(req.user, authz.actions.Read, req.context.narrative);
+
   if (!(await req.context.narrative.exists())) throw new NotFound();
   return next();
 };
@@ -321,6 +325,9 @@ function pathParts(path = "") {
 function sendSubresource(subresourceExtractor) {
   return async (req, res) => {
     const subresource = subresourceExtractor(req);
+
+    authz.assertAuthorized(req.user, authz.actions.Read, subresource.resource);
+
     const subresourceUrl = await subresource.url();
 
     /* Proxy the data through us:
@@ -369,6 +376,9 @@ function receiveSubresource(subresourceExtractor) {
   return async (req, res) => {
     const method = "PUT";
     const subresource = subresourceExtractor(req);
+
+    authz.assertAuthorized(req.user, authz.actions.Write, subresource.resource);
+
     const subresourceUrl = await subresource.url(method);
 
     /* Proxy the data through us:
